+++ src/components/hasIdAndProperties.js
