+++ src/readStream.js
@@ -7,7 +7,7 @@ var isValidId = require('./components/isValidId');
 var calculateFilePath = require('./components/calculateFilePath');
 var filterOutUnreadableFiles = require('./components/filterOutUnreadableFiles');
 var loadJSON = require('./components/loadJSON');
-var filterOutIncompleteRecords = require('./components/filterOutIncompleteRecords');
+var hasIdAndProperties = require('./components/hasIdAndProperties');
 var isNotDeprecatedRecord = require('./components/isNotDeprecatedRecord');
 var extractFields = require('./components/extractFields');
 var filterOutNamelessRecords = require('./components/filterOutNamelessRecords');
@@ -24,7 +24,7 @@ function readData(directory, types, wofRecords, callback) {
       .pipe(calculateFilePath.create())
       .pipe(filterOutUnreadableFiles.create(directory + 'data/'))
       .pipe(loadJSON.create(directory + 'data/'))
-      .pipe(filterOutIncompleteRecords.create())
+      .pipe(hasIdAndProperties.create())
       .pipe(isNotDeprecatedRecord.create())
       .pipe(extractFields.create())
       .pipe(filterOutNamelessRecords.create())
